Import latest changes to the bitmap library.
Main development is now going on in the Bigtop fork. This import includes the
following changes:

cl/68071490
Cache BitmapShader instance in CircularBitmapDrawable.
This simple implementation is eliminating almost all of the BitmapShader
construction during startup.

cl/69907083
Add compatibility mode to StyledCornersBitmapDrawable for b/15023700 in 4.4.3

The rounded corners and flaps are now drawn on the canvas, instead of clipped
with a path. This part is in the bitmap library.

The drawn corners must match the color of the background color of the
container. The bt_megalist_selected_item_background color has been pre-mixed
with @android:color/white so it's in a ColorDrawable instead of a
LayerDrawable, and so its color can be used to draw the fake corners. This part
is in the Bigtop codebase.

Change-Id: I21a22d8550fbe1dd3de7410cd82969ff947c27ea